Pass projects instead
