+++ frontend/src/admin/user-form.tsx
@@ -1,9 +1,9 @@
 import React, { FormEvent, FunctionComponent, useEffect, useState } from "react";
-import { useParams, useNavigate } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import { PrimeReactProvider } from "primereact/api";
 import { InputText } from "primereact/inputtext";
 import { Button } from "primereact/button";
-import { User } from "zero-zummon";
+import { User, Project, projectsFromJson } from "zero-zummon";
 import { redirectToLogin } from "./use-users";
 import { UserProjectsList } from "./user-projects-list";
 
@@ -11,10 +11,10 @@ export const UserForm: FunctionComponent = () => {
     const {userId} = useParams<{ userId: string }>();
     const [user, setUser] = useState<User | null>(null);
     const [originalData, setOriginalData] = useState<User | null>(null);
+    const [userProjects, setUserProjects] = useState<Project[]>([]);
 
     const [loading, setLoading] = useState(false);
     const [isEditing, setIsEditing] = useState(false);
-    const navigate = useNavigate();
 
     const handleCancel = () => {
         if (originalData) {
@@ -59,6 +59,21 @@ export const UserForm: FunctionComponent = () => {
                     setLoading(false);
                 }
             };
+
+            const fetchProjects = async () => {
+                try {
+                    const response = await fetch(`${import.meta.env.VITE_ZTOR_URL}/users/${userId}/projects`, {
+                        credentials: "include",
+                    });
+                    if (!response.ok) return
+    
+                    setUserProjects(projectsFromJson(await response.text()))
+                } catch (error) {
+                    console.error("Error fetching projects:", error);
+                }
+            };
+    
+            fetchProjects();
             fetchUser();
         } else {
             setIsEditing(true);
@@ -86,8 +101,6 @@ export const UserForm: FunctionComponent = () => {
             if (!response.ok) {
                 alert(`Error: ${response.statusText}`);
             }
-
-            navigate(`/users`);
         } finally {
             setIsEditing(false);
             setLoading(false);
@@ -143,7 +156,7 @@ export const UserForm: FunctionComponent = () => {
                         )}
                     </div>
                 </form>
-                <UserProjectsList userId={userId}/>
+                <UserProjectsList projects={userProjects}/>
             </div>
         </PrimeReactProvider>
     );
